Solution for day3part2
Similar to the solution for part1, this makes a small change to what we
store in the initial HashMap: rather than tracking how many claims
refer to a specific coordinate, we are instead storing the claim_id of
the first claim to reference that point.

This enables us to use a second HashMap that tracks
<claim_id: isize, no_overlap: bool>. As we parse each claim out, we
initialize the claim_id into this new map as having no overlap. Then,
as we walk through the coordinates it claims, we perform an update if
we find that a coordinate already had a reference. At the same time, we
also update the claim that is part of this collision as having overlap

At the end of this, we can walk to claim map to find any claims that
still have no overlap and report that out