Alternate approach to edge detection that catches a missed case
No idea _what_ was specifically wrong in the previous approach, but an
alternate way to find infinite area points is to grow the grid. Any
point that gains area during this process can be considered to be an
infinite area

To accomodate this, don't walk the two axes initially - this will give
us non-negative space to expand to in that direction. Additionally, we
can expand beyond the known-max lines as much as we want.